requirements.txt
# Core Pydantic AI Framework
pydantic-ai>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI Integration
openai>=1.0.0

# Database and Storage
supabase>=2.0.0
asyncpg>=0.28.0

# Web Framework (for API endpoints)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP and Async
httpx>=0.25.0
aiofiles>=23.0.0

# Environment Management
python-dotenv>=1.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Testing and Validation
pytest-mock>=3.12.0
httpx-mock>=0.10.0